Refactor card css to maintain height. Separate thumbnail to own component. Video Updates uses this
