+++ src/components/NewsCard/NewsCard.module.scss
@@ -4,10 +4,10 @@
   width: 100%;
 
   &.large {
-    gap: 10px;
+    gap: 5px;
     grid-template-areas: "title title" "caption caption" "date author";
     grid-template-columns: 1fr 1fr;
-    grid-template-rows: 1fr 1fr 1fr;
+    grid-template-rows: auto 40px 1fr;
   }
 
   &.small {
