+++ src/components/NewsThumb/NewsThumb.module.scss
@@ -1,0 +1,32 @@
+.NewsThumb {
+  align-items: center;
+  aspect-ratio: 16 / 9;
+  background-color: #eee;
+  border: 1px solid #ddd;
+  display: flex;
+  justify-content: center;
+  width: 100%;
+
+  & .Icon {
+    &.github {
+      height: 13%;
+    }
+
+    &.medium {
+      height: 18%;
+    }
+
+    &.discord {
+      height: 13%;
+    }
+
+    &.youtube {
+      height: 18%;
+    }
+
+    &.xangle {
+      fill: #4e33c2;
+      height: 13%;
+    }
+  }
+}
