+++ src/components/VideoCard/index.tsx
@@ -7,6 +7,7 @@ import VideoPlayer from "@/components/VideoPlayer";
 import { NewsDataObj } from "@/types/index";
 
 import styles from "./VideoCard.module.scss";
+import NewsThumb from "../NewsThumb";
 
 interface Props {
   newsData: NewsDataObj;
@@ -17,7 +18,11 @@ function VideoCard({ newsData, size }: Props) {
   return (
     <div className={classNames(styles.VideoCard, styles[size])}>
       <div className={classNames(styles.Video, styles[size])}>
-        <VideoPlayer url={newsData.url} />
+        {size === "large" ? (
+          <VideoPlayer url={newsData.url} />
+        ) : (
+          <NewsThumb source={newsData.source} thumbUrl={newsData.thumbnail} />
+        )}
       </div>
       <NewsCard newsData={newsData} size={size} />
     </div>
